fix or silence compiler warnings (msvc) about narrowing conversions, non-copyable types, dll-interfaces and trivias
(cherry picked from commit 3d54537e827db4a94206a4fd85fbe3816b18e83f)

+++ include/boost/iostreams/filter/test.hpp
@@ -101,7 +101,7 @@ public:
             return -1;
         std::streamsize avail = 
             (std::min) (n, static_cast<std::streamsize>(data_.size() - pos_));
-        std::streamsize amt = (std::min) (rand(inc_), avail);
+        std::streamsize amt = (std::min) (rand(static_cast<int>(inc_)), avail);
         if (amt)
             std::memcpy(s, data_.c_str() + pos_, amt);
         pos_ += amt;
@@ -129,7 +129,7 @@ public:
         { }
     std::streamsize write(const char* s, std::streamsize n)
     {
-        std::streamsize amt = (std::min) (rand(inc_), n);
+        std::streamsize amt = (std::min) (rand(static_cast<int>(inc_)), n);
         dest_.insert(dest_.end(), s, s + amt);
         return amt;
     }
